+++ aio2ch/__init__.py
@@ -1,5 +1,5 @@
 __author__ = 'wkpn'
-__version__ = '2.0.3'
+__version__ = '2.1.0'
 __license__ = 'MIT'
 __url__ = 'https://github.com/wkpn/aio2ch'