+++ aio2ch/helpers.py
@@ -34,3 +34,17 @@ def get_board_and_thread_from_url(thread_url: str) -> Union[Tuple[str, str], Tup
         return board, thread
 
     return None, None
+
+
+def clean_html_tags(raw_text: str) -> str:
+    if not raw_text:  # if no text e.g. empty string
+        return raw_text
+
+    clean_text = re.sub('<br>', '\n', raw_text)
+    clean_text = re.sub('&quot;', '"', clean_text)
+    clean_text = re.sub('&#47;', '/', clean_text)
+
+    tags_matcher = re.compile('<.*?>')
+    clean_text = re.sub(tags_matcher, '', clean_text)
+
+    return clean_text