+++ tests/conftest.py
@@ -55,3 +55,21 @@ def invalid_thread_url():
 @pytest.fixture
 def number_of_threads(number=5):
     return number
+
+
+@pytest.fixture
+def raw_text():
+    return 'КУКУ ТАМ ОНЛИФЕНС СЛИЛИ, ЕСТЬ ЧТО У КОГО?<br>' \
+           '<a href=\"https:&#47;&#47;tjournal.ru&#47;internet&#47;146526-v-set-popali-terabayty-' \
+           'dannyh-onlyfans-sayta-gde-pornozvezdy-i-modeli-razmeshchayut-platnye-eroticheskie-foto-i-video\" ' \
+           'target=\"_blank\" rel=\"nofollow noopener noreferrer\">https:&#47;&#47;tjournal.ru&#47;internet&#47;' \
+           '146526-v-set-popali-terabayty-dannyh-onlyfans-sayta-gde-pornozvezdy-i-modeli-razmeshchayut-platnye-' \
+           'eroticheskie-foto-i-video</a><br><br>Давайте братья скажите где скачать эти терабайты'
+
+
+@pytest.fixture
+def clean_text():
+    return 'КУКУ ТАМ ОНЛИФЕНС СЛИЛИ, ЕСТЬ ЧТО У КОГО?\n' \
+           'https://tjournal.ru/internet/146526-v-set-popali-terabayty-dannyh-onlyfans-' \
+           'sayta-gde-pornozvezdy-i-modeli-razmeshchayut-platnye-eroticheskie-foto-i-video\n\n' \
+           'Давайте братья скажите где скачать эти терабайты'